Move some code to a shared location so jsdom15 can reuse it
