+++ src/gateway/environments/jsdom-sixteen/__tests__/jsdom-sixteen-resource-loader.test.js
@@ -2,7 +2,7 @@
 import * as JSDOM from "jsdom16";
 
 import * as Request from "../../../request.js";
-import * as ApplyAbortablePromisesPatch from "../apply-abortable-promises-patch.js";
+import * as ApplyAbortablePromisesPatch from "../../shared/apply-abortable-promises-patch.js";
 import * as Shared from "../../../../shared/index.js";
 
 import {JSDOMSixteenResourceLoader} from "../jsdom-sixteen-resource-loader.js";
@@ -10,7 +10,7 @@ import {JSDOMSixteenResourceLoader} from "../jsdom-sixteen-resource-loader.js";
 jest.mock("jsdom16");
 jest.mock("../../../../shared/index.js");
 jest.mock("../../../request.js");
-jest.mock("../apply-abortable-promises-patch.js");
+jest.mock("../../shared/apply-abortable-promises-patch.js");
 
 describe("JSDOMSixteenResourceLoader", () => {
     describe("#constructor", () => {
