+++ src/gateway/environments/jsdom-sixteen/jsdom-sixteen-file-resource-loader.js
@@ -4,7 +4,7 @@ import path from "path";
 import {promisify} from "util";
 import {ResourceLoader} from "jsdom16";
 import type {FetchOptions} from "jsdom16";
-import {applyAbortablePromisesPatch} from "./apply-abortable-promises-patch.js";
+import {applyAbortablePromisesPatch} from "../shared/apply-abortable-promises-patch.js";
 
 const readFileAsync = promisify(fs.readFile);
