+++ src/gateway/environments/jsdom-sixteen/types.js
@@ -1,45 +1,7 @@
 // @flow
 import type {ResourceLoader} from "jsdom16";
 import type {RenderAPI} from "../../types.js";
-
-/**
- * A thing that can be closed.
- */
-export interface ICloseable {
-    /**
-     * Close the closeable.
-     */
-    +close?: () => void;
-}
-
-/**
- * Gate API for control flow.
- */
-export interface IGate extends ICloseable {
-    /**
-     * Open the gate.
-     */
-    open(): void;
-
-    /**
-     * Close the gate.
-     */
-    close(): void;
-
-    /**
-     * True, if the gate is open; otherwise, false.
-     */
-    get isOpen(): boolean;
-}
-
-/**
- * Standard timer API as implemented by Node's global or a browser window.
- */
-export interface ITimerAPI {
-    setTimeout: $PropertyType<window, "setTimeout">;
-    setInterval: $PropertyType<window, "setInterval">;
-    requestAnimationFrame: $PropertyType<window, "requestAnimationFrame">;
-}
+import type {ICloseable} from "../shared/types.js";
 
 /**
  * A resource loader for use with JSDOM that can also have a close method for
