+++ src/gateway/environments/shared/__tests__/patch-against-dangling-timers.test.js
