+++ src/gateway/environments/shared/apply-abortable-promises-patch.js
