+++ src/gateway/environments/shared/patch-against-dangling-timers.js
